BUILDSYS-322 separated validation from the core snacc.c file
