+++ compiler/core/snacc-validators.h
@@ -1,0 +1,5 @@
+
+#include "../../c-lib/include/asn-incl.h"
+#include "asn1module.h"
+
+void ValidateASN1Data(ModuleList* allMods);
