+++ compiler/core/snacc.c
@@ -65,6 +65,7 @@ char* bVDAGlobalDLLExport = (char*)0;
 #include "print.h"
 #include "define.h"
 #include "snacc-util.h"
+#include "snacc-validators.h"
 #include "filetype.h"
 #include "../back-ends/structure-util.h"
 #include "../back-ends/str-util.h"
@@ -115,7 +116,6 @@ void PrintCCode PROTO((FILE * src, FILE* hdr, ModuleList* mods, Module* m, CRule
 void PrintIDLCode PROTO((FILE * idl, ModuleList* mods, Module* m, IDLRules* r, long int longJmpVal, int printValues));
 void ProcessMacros PROTO((Module * m));
 void SortAllDependencies PROTO((ModuleList * m));
-void ValidateStructure PROTO((ModuleList * m));
 int yyparse();
 
 /* Internal routines */
@@ -182,10 +182,8 @@ int gPrivateSymbols = 1;
 // Any deprecated information from a sequence, attribute or operation comment lower or equal than the gi64NoDeprecatedSymbols will get removed
 long long gi64NoDeprecatedSymbols = 0;
 
-// jan 11.1.2023 - Default level for validating the content of the asn1 files
-//  1 - Validates that operationIDs are not used twice
-//  2 - Validates that the result is an AsnRequestError and argument, result, and error are SEQUENCES and thus extensible (@deprecated modules are excluded from that check)
-int giValidationLevel = 2;
+// jan 11.1.2023 - Default level for validating the content of the asn1 files is all checks that the tool knows
+int giValidationLevel = 99999;
 
 // Write comments to the target files on true (parsing is always enabled)
 int giWriteComments = 0;
@@ -242,10 +240,13 @@ void Usage PARAMS((prgName, fp), char* prgName _AND_ FILE* fp)
 	fprintf(fp, "  -noprivate   do not generate code that is marked as private\n");
 	fprintf(fp, "  -nodeprecated   do not generate code that is marked as deprecated (any date)\n");
 	fprintf(fp, "  -nodeprecated:[Day.Month.Year]  do not generate code that has been marked deprecated prior to this date\n");
-	fprintf(fp, "  -ValidationLevel [0-2]   validate the asn1 against predefined rules\n");
+	fprintf(fp, "  -ValidationLevel n - Sets a specific validation rule set for the asn1 files. Default is that all of the following checks are applied\n");
 	fprintf(fp, "   0 no validation\n");
 	fprintf(fp, "   1 Validates that operationIDs are not used twice\n");
-	fprintf(fp, "   2 Validates that the result is an AsnRequestError and argument, result, and error are SEQUENCES and thus extensible (@deprecated modules are excluded from that check)\n");
+	fprintf(fp, "   2 Validates that operation arguments, results and errors are sequences or choices (only types are extendable) (@deprecated are not validated)\n");
+	fprintf(fp, "   4 Validates that errors are of the same type to generalize error handling (@deprecated are not validated)\n");
+	fprintf(fp, "   8 Validates that all sequences and choices contain ... to allow extending them (@deprecated are not validated)\n");
+	fprintf(fp, "   16 Validates that only allow types from the esnacc_whiteliste.json are used (@deprecated are not validated)\n");
 	fprintf(fp, "  -h   prints this msg\n");
 	fprintf(fp, "  -P   print the parsed ASN.1 modules to stdout from their parse trees\n");
 	fprintf(fp, "       (helpful debugging)\n");
@@ -1009,7 +1010,7 @@ int main PARAMS((argc, argv), int argc _AND_ char** argv)
 	 * Validates that the structures written are okay
 	 * Checks for certain things we do not want to see :)
 	 */
-	ValidateStructure(allMods);
+	ValidateASN1Data(allMods);
 
 	/*
 	 * STEP 12
@@ -2414,423 +2415,6 @@ int ModNamesUnique PARAMS((mods), ModuleList* mods)
 	return retVal;
 } /* ModNamesUnique */
 
-void EnsureNoDuplicateMethodIDs(ModuleList* allMods)
-{
-	Module* currMod;
-
-	const int MALLOC_SIZE = 50000;
-	int* ids = malloc(MALLOC_SIZE);
-	if (!ids)
-	{
-		snacc_exit("Out of memory");
-		return;
-	}
-	memset(ids, 0x00, MALLOC_SIZE);
-	int counter = 0;
-	int nWeHaveErrors = 0;
-
-	FOR_EACH_LIST_ELMT(currMod, allMods)
-	{
-		if (currMod->ImportedFlag == FALSE)
-		{
-			if (HasROSEOperations(currMod))
-			{
-				ValueDef* vd;
-				FOR_EACH_LIST_ELMT(vd, currMod->valueDefs)
-				{
-					if (!IsROSEValueDef(currMod, vd))
-						continue;
-
-					int methodID = vd->value->basicValue->a.integer;
-					for (int iCount = 0; iCount < counter; iCount++)
-					{
-						if (ids[iCount] == methodID)
-						{
-							fprintf(stderr, "Method/Event ID %i has been used multiple times.\n", methodID);
-							nWeHaveErrors = 1;
-							break;
-						}
-					}
-
-					ids[counter] = methodID;
-					counter++;
-				}
-			}
-		}
-	}
-	if (nWeHaveErrors)
-		snacc_exit_now(__FUNCTION__, "\nYou must ensure that the method/event IDs are only used once!\nNow terminating...\n");
-
-	free(ids);
-}
-
-void EnsureNoSequenceAndSetOfInArgumentOrResult(ModuleList* allMods)
-{
-	Module* currMod;
-	int nWeHaveErrors = 0;
-	const char* szLastErrorFile = NULL;
-	FOR_EACH_LIST_ELMT(currMod, allMods)
-	{
-		if (currMod->ImportedFlag == FALSE)
-		{
-			if (HasROSEOperations(currMod))
-			{
-				ValueDef* vd;
-				FOR_EACH_LIST_ELMT(vd, currMod->valueDefs)
-				{
-					if (!IsROSEValueDef(currMod, vd))
-						continue;
-
-					if (IsDeprecatedFlaggedOperation(currMod, vd->definedName))
-						continue;
-
-					char* pszArgument = NULL;
-					char* pszResult = NULL;
-					char* pszError = NULL;
-					Type* argumentType = NULL;
-					Type* resultType = NULL;
-					Type* errorType = NULL;
-					if (GetROSEDetails(currMod, vd, &pszArgument, &pszResult, &pszError, &argumentType, &resultType, &errorType, false))
-					{
-						struct BasicType* argumentBasicType = NULL;
-						bool bArgumentIssue = false;
-						if (argumentType)
-						{
-							argumentBasicType = ResolveBasicTypeReferences(argumentType->basicType, NULL);
-							if (!argumentBasicType)
-								bArgumentIssue = true;
-							else
-							{
-								enum BasicTypeChoiceId choiceId = argumentBasicType->choiceId;
-								if (choiceId != BASICTYPE_SEQUENCE && choiceId != BASICTYPE_CHOICE)
-									bArgumentIssue = true;
-							}
-						}
-						struct BasicType* resultBasicType = NULL;
-						bool bResultIssue = false;
-						if (resultType)
-						{
-							resultBasicType = ResolveBasicTypeReferences(resultType->basicType, NULL);
-							if (!resultBasicType)
-								bResultIssue = true;
-							else
-							{
-								enum BasicTypeChoiceId choiceId = resultBasicType->choiceId;
-								if (choiceId != BASICTYPE_SEQUENCE && choiceId != BASICTYPE_CHOICE)
-									bResultIssue = true;
-							}
-						}
-
-						bool bErrorIssue = false;
-						bool bWrongErrorObject = false;
-						if (errorType)
-						{
-							bErrorIssue = errorType->basicType->choiceId != BASICTYPE_SEQUENCE;
-							bWrongErrorObject = pszError && strcmp(pszError, "AsnRequestError") != 0;
-						}
-						if (bArgumentIssue || bResultIssue || bErrorIssue || bWrongErrorObject)
-						{
-							if (szLastErrorFile != currMod->asn1SrcFileName)
-							{
-								if (szLastErrorFile)
-									fprintf(stderr, "\n");
-#ifdef _WIN32
-								HANDLE hErr = GetStdHandle(STD_ERROR_HANDLE);
-								CONSOLE_SCREEN_BUFFER_INFO csbiInfo;
-								GetConsoleScreenBufferInfo(hErr, &csbiInfo);
-								SetConsoleTextAttribute(hErr, FOREGROUND_RED | FOREGROUND_INTENSITY);
-#endif // _WIN32
-								fprintf(stderr, "Errors in %s:\n", currMod->asn1SrcFileName);
-#ifdef _WIN32
-								SetConsoleTextAttribute(hErr, FOREGROUND_RED | FOREGROUND_GREEN | FOREGROUND_BLUE);
-#endif // _WIN32
-								szLastErrorFile = currMod->asn1SrcFileName;
-							}
-							if (bArgumentIssue)
-							{
-								fprintf(stderr, "- %s is using %s as argument which is a ", vd->definedName, pszArgument);
-								PrintTypeById(stderr, argumentType->basicType->choiceId);
-								fprintf(stderr, ".\n  You must use a SEQUENCE or CHOICE here (expandability).\n");
-								nWeHaveErrors++;
-							}
-							if (bResultIssue)
-							{
-								fprintf(stderr, "- %s is using %s as result which is a ", vd->definedName, pszResult);
-								PrintTypeById(stderr, resultType->basicType->choiceId);
-								fprintf(stderr, ".\n  You must use a SEQUENCE or CHOICE here (expandability).\n");
-								nWeHaveErrors++;
-							}
-							if (bErrorIssue)
-							{
-								fprintf(stderr, "- %s is using %s as error which is a ", vd->definedName, pszError);
-								PrintTypeById(stderr, errorType->basicType->choiceId);
-								fprintf(stderr, ".\n  You must use a SEQUENCE here (expandability).\n");
-								nWeHaveErrors++;
-							}
-							if (bWrongErrorObject)
-							{
-								fprintf(stderr, "- %s is using %s as error but must use AsnRequestError.\n", vd->definedName, pszError);
-								nWeHaveErrors++;
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-	if (nWeHaveErrors)
-	{
-		fprintf(stderr, "\n");
-		fprintf(stderr, "*************************************************************\n");
-		fprintf(stderr, "* Methods may contain issues if they are flagged deprecated *\n");
-		fprintf(stderr, "*************************************************************\n\n");
-		fprintf(stderr, "- found %i errors\n", nWeHaveErrors);
-		snacc_exit_now(__FUNCTION__, "Now terminating...\n");
-	}
-}
-
-const char* getTypeName(enum BasicTypeChoiceId choiceId)
-{
-	switch (choiceId)
-	{
-		case BASICTYPE_BOOLEAN:
-			return "BASICTYPE_BOOLEAN";
-		case BASICTYPE_INTEGER:
-			return "BASICTYPE_INTEGER";
-		case BASICTYPE_BITSTRING:
-			return "BASICTYPE_BITSTRING";
-		case BASICTYPE_OCTETSTRING:
-			return "BASICTYPE_OCTETSTRING";
-		case BASICTYPE_NULL:
-			return "BASICTYPE_NULL";
-		case BASICTYPE_OID:
-			return "BASICTYPE_OID";
-		case BASICTYPE_REAL:
-			return "BASICTYPE_REAL";
-		case BASICTYPE_ENUMERATED:
-			return "BASICTYPE_ENUMERATED";
-		case BASICTYPE_SEQUENCE:
-			return "BASICTYPE_SEQUENCE";
-		case BASICTYPE_SEQUENCEOF:
-			return "BASICTYPE_SEQUENCEOF";
-		case BASICTYPE_SET:
-			return "BASICTYPE_SET";
-		case BASICTYPE_SETOF:
-			return "BASICTYPE_SETOF";
-		case BASICTYPE_CHOICE:
-			return "BASICTYPE_CHOICE";
-		case BASICTYPE_SELECTION:
-			return "BASICTYPE_SELECTION";
-		case BASICTYPE_COMPONENTSOF:
-			return "BASICTYPE_COMPONENTSOF";
-		case BASICTYPE_ANY:
-			return "BASICTYPE_ANY";
-		case BASICTYPE_ANYDEFINEDBY:
-			return "BASICTYPE_ANYDEFINEDBY";
-		case BASICTYPE_LOCALTYPEREF:
-			return "BASICTYPE_LOCALTYPEREF";
-		case BASICTYPE_IMPORTTYPEREF:
-			return "BASICTYPE_IMPORTTYPEREF";
-		case BASICTYPE_MACROTYPE:
-			return "BASICTYPE_MACROTYPE";
-		case BASICTYPE_MACRODEF:
-			return "BASICTYPE_MACRODEF";
-		case BASICTYPE_NUMERIC_STR:
-			return "BASICTYPE_NUMERIC_STR";
-		case BASICTYPE_PRINTABLE_STR:
-			return "BASICTYPE_PRINTABLE_STR";
-		case BASICTYPE_UNIVERSAL_STR:
-			return "BASICTYPE_UNIVERSAL_STR";
-		case BASICTYPE_IA5_STR:
-			return "BASICTYPE_IA5_STR";
-		case BASICTYPE_BMP_STR:
-			return "BASICTYPE_BMP_STR";
-		case BASICTYPE_UTF8_STR:
-			return "BASICTYPE_UTF8_STR";
-		case BASICTYPE_UTCTIME:
-			return "BASICTYPE_UTCTIME";
-		case BASICTYPE_GENERALIZEDTIME:
-			return "BASICTYPE_GENERALIZEDTIME";
-		case BASICTYPE_GRAPHIC_STR:
-			return "BASICTYPE_GRAPHIC_STR";
-		case BASICTYPE_VISIBLE_STR:
-			return "BASICTYPE_VISIBLE_STR";
-		case BASICTYPE_GENERAL_STR:
-			return "BASICTYPE_GENERAL_STR";
-		case BASICTYPE_OBJECTDESCRIPTOR:
-			return "BASICTYPE_OBJECTDESCRIPTOR";
-		case BASICTYPE_VIDEOTEX_STR:
-			return "BASICTYPE_VIDEOTEX_STR";
-		case BASICTYPE_T61_STR:
-			return "BASICTYPE_T61_STR";
-		case BASICTYPE_EXTERNAL:
-			return "BASICTYPE_EXTERNAL";
-		case BASICTYPE_OCTETCONTAINING:
-			return "BASICTYPE_OCTETCONTAINING";
-		case BASICTYPE_BITCONTAINING:
-			return "BASICTYPE_BITCONTAINING";
-		case BASICTYPE_RELATIVE_OID:
-			return "BASICTYPE_RELATIVE_OID";
-		case BASICTYPE_EXTENSION:
-			return "BASICTYPE_EXTENSION";
-		case BASICTYPE_SEQUENCET:
-			return "BASICTYPE_SEQUENCET";
-		case BASICTYPE_OBJECTCLASS:
-			return "BASICTYPE_OBJECTCLASS";
-		case BASICTYPE_OBJECTCLASSFIELDTYPE:
-			return "BASICTYPE_OBJECTCLASSFIELDTYPE";
-		case BASICTYPE_ASNSYSTEMTIME:
-			return "BASICTYPE_ASNSYSTEMTIME";
-		default:
-			assert(FALSE);
-			return "UNKNOWN";
-	}
-}
-
-bool isSupportedType(enum BasicTypeChoiceId choiceId)
-{
-	switch (choiceId)
-	{
-		case BASICTYPE_BOOLEAN:
-		case BASICTYPE_INTEGER:
-		case BASICTYPE_OCTETSTRING:
-		case BASICTYPE_ENUMERATED:
-		case BASICTYPE_SEQUENCE:
-		case BASICTYPE_SEQUENCEOF:
-		case BASICTYPE_CHOICE:
-		case BASICTYPE_REAL:
-		case BASICTYPE_UTF8_STR:
-		case BASICTYPE_EXTENSION:
-		case BASICTYPE_NULL:
-		case BASICTYPE_ANY:
-			// We support these tags
-			return true;
-		case BASICTYPE_LOCALTYPEREF:
-		case BASICTYPE_IMPORTTYPEREF:
-			// Each import should be resolved BEFORE calling this method
-			assert(FALSE);
-			return false;
-		default:
-			return false;
-	}
-}
-
-// Returns true when an invalid element was found
-bool recurseFindInvalid(Module* mod, Type* type, const char* szPath, const char* szElementName)
-{
-#define TESTBUFFERSIZE 256
-#define BUFFERSIZE 4096
-
-	bool bFoundInvalid = false;
-	char szCurrentPath[BUFFERSIZE] = {0};
-	strcpy_s(szCurrentPath, BUFFERSIZE, szPath);
-
-	enum BasicTypeChoiceId choiceId = type->basicType->choiceId;
-
-	if (szElementName)
-	{
-		if ((choiceId == BASICTYPE_SEQUENCE || choiceId == BASICTYPE_BITSTRING) && IsDeprecatedFlaggedSequence(mod, szElementName))
-			return false;
-		char szNewName[TESTBUFFERSIZE] = {0};
-		strcat_s(szNewName, TESTBUFFERSIZE, "::");
-		strcat_s(szNewName, TESTBUFFERSIZE, szElementName);
-		if ((choiceId == BASICTYPE_SEQUENCE || choiceId == BASICTYPE_LOCALTYPEREF || choiceId == BASICTYPE_IMPORTTYPEREF) && type->cxxTypeRefInfo->className)
-		{
-			if (IsDeprecatedFlaggedSequence(mod, type->cxxTypeRefInfo->className))
-				return false;
-			strcat_s(szNewName, TESTBUFFERSIZE, "(");
-			strcat_s(szNewName, TESTBUFFERSIZE, type->cxxTypeRefInfo->className);
-			strcat_s(szNewName, TESTBUFFERSIZE, ")");
-		}
-
-		char szTest2[TESTBUFFERSIZE] = {0};
-		strcat_s(szTest2, TESTBUFFERSIZE, szNewName);
-		strcat_s(szTest2, TESTBUFFERSIZE, "::");
-		if (strstr(szPath, szTest2))
-		{
-			// Recursion check -> haben wir schon (raus...)
-			return false;
-		}
-
-		const char* pos = strstr(szPath, szNewName);
-		if (pos)
-		{
-			// Is it at the end of the string?
-			const size_t len1 = strlen(pos);
-			const size_t len2 = strlen(szNewName);
-			if (len1 == len2)
-			{
-				// Recursion check -> haben wir schon (raus...)
-				return false;
-			}
-		}
-
-		strcat_s(szCurrentPath, BUFFERSIZE, szNewName);
-	}
-
-	if (choiceId == BASICTYPE_LOCALTYPEREF)
-		bFoundInvalid = recurseFindInvalid(mod, type->basicType->a.localTypeRef->link->type, szCurrentPath, NULL);
-	else if (choiceId == BASICTYPE_IMPORTTYPEREF)
-		bFoundInvalid = recurseFindInvalid(mod, type->basicType->a.importTypeRef->link->type, szCurrentPath, NULL);
-	else
-	{
-		if (!isSupportedType(choiceId))
-		{
-			fprintf(stderr, "Unsupported type %s found in %s\n", getTypeName(choiceId), szCurrentPath);
-			return true;
-		}
-
-		if (choiceId == BASICTYPE_SEQUENCEOF)
-		{
-			Type* subType = type->basicType->a.sequenceOf;
-			bFoundInvalid = recurseFindInvalid(mod, subType, szCurrentPath, NULL);
-		}
-		else if (choiceId == BASICTYPE_SEQUENCE)
-		{
-			NamedTypeList* typeList = type->basicType->a.sequence;
-			NamedType* subType;
-			FOR_EACH_LIST_ELMT(subType, typeList)
-			{
-				// Due to possible recursion we need to store the current position
-				AsnListNode* oldCurr = typeList->curr;
-				if (recurseFindInvalid(mod, subType->type, szCurrentPath, subType->fieldName))
-					bFoundInvalid = true;
-				typeList->curr = oldCurr;
-			}
-		}
-	}
-	return bFoundInvalid;
-}
-
-void EnsureOnlySupportedObjects(ModuleList* allMods)
-{
-	Module* currMod;
-	int nWeHaveErrors = 0;
-	FOR_EACH_LIST_ELMT(currMod, allMods)
-	{
-		if (currMod->ImportedFlag == FALSE)
-		{
-			TypeDef* vd;
-			FOR_EACH_LIST_ELMT(vd, currMod->typeDefs)
-			{
-				char szPath[128] = {0};
-				sprintf_s(szPath, 128, "%s ", currMod->asn1SrcFileName);
-				if (recurseFindInvalid(currMod, vd->type, szPath, vd->definedName))
-					nWeHaveErrors++;
-			}
-		}
-	}
-
-	if (nWeHaveErrors)
-	{
-		fprintf(stderr, "\n");
-		fprintf(stderr, "**********************************\n");
-		fprintf(stderr, "* Found not supported asn1 types *\n");
-		fprintf(stderr, "**********************************\n");
-		snacc_exit_now(__FUNCTION__, "Now terminating...\n");
-	}
-}
-
 void CreateNames(ModuleList* allMods)
 {
 	Module* currMod;
@@ -2856,17 +2440,6 @@ void CreateNames(ModuleList* allMods)
 		currMod->ROSEHdrForwardDeclFileName = MakeROSEHdrForwardDeclFileName(currMod->baseFileName);
 		currMod->ROSESrcJAVAFileName = MakeROSESrcJAVAFileName(currMod->baseFileName);
 		currMod->ROSESwiftInterfaceFileName = MakeROSESwiftInterfaceFileName(currMod->baseFileName);
-	}
-}
-
-void ValidateStructure(ModuleList* allMods)
-{
-	if (giValidationLevel >= 1)
-		EnsureNoDuplicateMethodIDs(allMods);
-	if (giValidationLevel >= 2)
-	{
-		EnsureOnlySupportedObjects(allMods);
-		EnsureNoSequenceAndSetOfInArgumentOrResult(allMods);
 	}
 }
